LayoutTests/fast/css3-text/font-synthesis-parse-expected.txt

PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 3
PASS declaration.getPropertyValue('font-synthesis') is "initial"
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 3
PASS declaration.getPropertyValue('font-synthesis') is "weight"
PASS computedStyle.getPropertyValue('font-synthesis') is "weight"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 3
PASS declaration.getPropertyValue('font-synthesis') is "style"
PASS computedStyle.getPropertyValue('font-synthesis') is "style"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 3
PASS declaration.getPropertyValue('font-synthesis') is "weight style"
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 3
PASS declaration.getPropertyValue('font-synthesis') is "weight style"
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
PASS declaration.getPropertyValue('font-synthesis') is ""
PASS computedStyle.getPropertyValue('font-synthesis') is "weight style small-caps"
PASS successfullyParsed is true

TEST COMPLETE
